chore: separate component
