+++ packages/wallets/src/components/WalletAddedSuccess/index.ts
@@ -1,0 +1,1 @@
+export { default as WalletAddedSuccess } from './WalletAddedSuccess';
